+++ src/database/operations/sources/index.ts
@@ -3,3 +3,4 @@ export * from './loadSource';
 export * from './loadSourceBundle';
 export * from './loadSystemSource';
 export * from './removeSource';
+export * from './updateSource';